requirements.txt
aiohttp>=3.9
requests>=2.31
python-dotenv>=1.0
discord.py>=2.3
